requirements.txt
Flask
flask-cors
ultralytics
Pillow
